requirements.txt
boto3==1.22.2
botocore==1.25.2
jmespath==1.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
